Fixes on the Product Model
Product's property Buyer had to be made nullable as well, as it is part of a foreign key pointing to the Users table. Also had to add a OnDelete to the two foreign keys BuyerID and SellerId in the DbContext's FLUENT API. Once those fixes were applied the datasets from the json files were able to be inserted in the database on the SQL server.